refactor : MarketError 추가
상황에 따른 오류 추가
